refactor: Move VehicleImage to vehicles folder
