+++ src/components/vehicles/VehicleImage.js
